server/requirements.txt
captcha==0.2.1
coreapi==2.1.1
Django==1.10.1
django-filter==1.0.1
django-redis-sessions==0.5.6
django-rest-swagger==2.1.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
djoser==0.5.1
hiredis==0.2.0
itypes==1.1.0
openapi-codec==1.2.1
Pillow==3.4.2
psycopg2==2.6.2
PyJWT==1.4.2
redis==2.10.5
requests==2.12.4
simplejson==3.10.0
tornado==4.4.1
tornado-redis==2.4.18
uritemplate==3.0.0
